ignore Crashpad folder

+++ tests/osn-tests/util/general.ts
@@ -80,7 +80,7 @@ export function deleteConfigFiles(): void {
         let currentFile: string;
         let files = fs.readdirSync(configFolderPath);
         files.forEach(file => {
-            if (file !== 'node-obs') {
+            if (file !== 'node-obs' && file !== 'Crashpad') {
                 currentFile = file;
                 try {
                     fs.unlinkSync(path.join(configFolderPath, file));
